+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ## 1.YY.Z (2023-MM-DD)
 
+- **ENH: Adding NODATA formalizations in `rasters(_rio)`: function to get the value from a dtype (`get_nodata_value` and global variables for `uint8`, `int8`, `uint16`, `float`)**
 - CI: Don't run tests when only `__init__` or `__meta__` is updated
 - CI: Move data elsewhere than git